Etteranmelding på Renn admin siden #109

+++ result_service_gui/services/contestants_adapter.py
@@ -9,9 +9,8 @@ from aiohttp import ClientSession
 from aiohttp import hdrs, web
 from multidict import MultiDict
 
-from .raceclasses_adapter import (
-    RaceclassesAdapter,
-)
+from .raceclasses_adapter import RaceclassesAdapter
+from .start_adapter import StartAdapter
 
 EVENTS_HOST_SERVER = os.getenv("EVENTS_HOST_SERVER", "localhost")
 EVENTS_HOST_PORT = os.getenv("EVENTS_HOST_PORT", "8082")
@@ -108,7 +107,7 @@ class ContestantsAdapter:
                 res = resp.status
                 logging.info(f"result - got response {res} - {resp}")
                 if res == 200:
-                    res = await resp.json()
+                    body = await resp.json()
                 elif resp.status == 401:
                     raise web.HTTPBadRequest(reason=f"401 Unathorized - {servicename}")
                 else:
@@ -117,7 +116,22 @@ class ContestantsAdapter:
                     raise web.HTTPBadRequest(
                         reason=f"Error - {resp.status}: {body['detail']}."
                     )
-        return str(res)
+        # trying to parse result - skip if it fails
+        informasjon = ""
+        try:
+            informasjon = f"Opprettet {body['created']} av {body['total']} deltakere."
+            if body["updated"]:
+                informasjon += f"<br><br>Duplikater ({len(body['updated'])}):"
+                for update in body["updated"]:
+                    informasjon += f"<br>- {update}"
+            if body["failures"]:
+                informasjon += f"<br><br>Error ({len(body['failures'])}):"
+                for failure in body["failures"]:
+                    informasjon += f"<br>- {failure}"
+        except Exception:
+            logging.error(f"Error parsing result {body}")
+
+        return informasjon
 
     async def delete_all_contestants(self, token: str, event_id: str) -> str:
         """Delete all contestants in one event function."""
@@ -146,17 +160,29 @@ class ContestantsAdapter:
         return str(res)
 
     async def delete_contestant(
-        self, token: str, event_id: str, contestant_id: str
+        self, token: str, event_id: str, contestant: dict
     ) -> str:
         """Delete one contestant function."""
         servicename = "delete_contestant"
+
+        # validation - if racer is in start-list, deletion not allowed
+        current_contestant = await ContestantsAdapter().get_contestant(
+            token, event_id, contestant["id"]
+        )
+        start_entries = await StartAdapter().get_start_entries_by_bib(
+            token, event_id, current_contestant["bib"]
+        )
+        if start_entries:
+            raise web.HTTPBadRequest(
+                reason=f"Startnr {current_contestant['bib']} kan ikke slettes fordi løper er i startliste."
+            )
+
         headers = {
             hdrs.AUTHORIZATION: f"Bearer {token}",
         }
-
         async with ClientSession() as session:
             async with session.delete(
-                f"{EVENT_SERVICE_URL}/events/{event_id}/contestants/{contestant_id}",
+                f"{EVENT_SERVICE_URL}/events/{event_id}/contestants/{contestant['id']}",
                 headers=headers,
             ) as resp:
                 res = resp.status
@@ -171,6 +197,21 @@ class ContestantsAdapter:
                     raise web.HTTPBadRequest(
                         reason=f"Error - {resp.status}: {body['detail']}."
                     )
+        # update number of contestants in raceclass
+        try:
+            klasse = await RaceclassesAdapter().get_raceclass_by_ageclass(
+                token, event_id, contestant["ageclass"]
+            )
+            if klasse:
+                klasse["no_of_contestants"] = klasse["no_of_contestants"] - 1
+            result = await RaceclassesAdapter().update_raceclass(
+                token, event_id, klasse["id"], klasse
+            )
+            logging.debug(f"No_of_contestants updated - {result}")
+        except Exception as e:
+            logging.error(
+                f"{servicename} failed on update no of contestants in raceclass {e} - {contestant['ageclass']}"
+            )
         return str(res)
 
     async def get_all_contestants(self, token: str, event_id: str) -> List:
@@ -354,6 +395,18 @@ class ContestantsAdapter:
         servicename = "update_contestant"
         request_body = copy.deepcopy(contestant)
         logging.debug(f"update_contestants, got request_body {request_body}")
+
+        # validation - if racer is in start-list, no changes are allowed
+        current_contestant = await ContestantsAdapter().get_contestant(
+            token, event_id, contestant["id"]
+        )
+        start_entries = await StartAdapter().get_start_entries_by_bib(
+            token, event_id, current_contestant["bib"]
+        )
+        if start_entries:
+            raise web.HTTPBadRequest(
+                reason=f"Startnr {current_contestant['bib']} kan ikke endres fordi løper er i startliste."
+            )
 
         url = f"{EVENT_SERVICE_URL}/events/{event_id}/contestants/{contestant['id']}"
         headers = MultiDict(
